third_party/blink/web_tests/virtual/http-cache-default-backend/external/wpt/fetch/http-cache/cache-mode.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] Fetch sends Cache-Control: no-cache and Pragma: no-cache when cache mode is no-store
  assert_equals: request 1 header cache-control value is "undefined", not "no-cache" expected (string) "no-cache" but got (undefined) undefined
Harness: the test ran to completion.
